backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
supabase>=2.10.0
anthropic>=0.39.0
openai>=1.55.0
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
pgvector>=0.3.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
